rpsで速度出せた

+++ Software/F446RE_line/lib/MovingAve/MovingAve.cpp
@@ -15,7 +15,7 @@ void MovingAve::Compute(float input) {
       cnt_++;
 }
 
-float MovingAve::Get() {
+double MovingAve::Get() {
       return result_;
 }
